Create BitcoinUtilsLegacy class to have deprecated methods. Make methods private when possible. Refactor tests class
