+++ rskj-core/src/test/java/co/rsk/peg/bitcoin/BitcoinUtilsLegacyTest.java
@@ -1,0 +1,4 @@
+package co.rsk.peg.bitcoin;
+
+public class BitcoinUtilsLegacyTest {
+}
